Add files via upload
A number of audio files no requiring high fidelity (such as general audio files) are often mistakenly produced at high bit rates than needed. I converted a 128KB down to 32KB and did not notice any difference in the hearing of the two. And that alone shrunk the file size down to 1/4 of original size. NOTE: The originals are left untouched. You select by checkbox 64, 48, or 32 or any two or all of them. They go into their own sub-folders. The program first checks by running ffmpeg (REQUIRED for the program to process the Mp3s) in the background with the help command text going into a temporary text file. If that file is over 200 bytes, then ffmpeg is certain to have been installed.